Move win module to platform namespace
