+++ salt/platform/win.py
